Estruturação ainda maior da interação com usuário. Correções pontuais de itens necessários para as interações e em combates

+++ tarefa4/src/com/rpg/combate/Combatente.java
@@ -16,7 +16,7 @@ public interface Combatente {
     int getPontosDeVida();
     int getPontosDeVidaMax();
     boolean estaVivo();
-    int receberDano(double dano);
+    double receberDano(double dano);
     void receberCura(int cura);
     AcaoDeCombate escolherAcao(Combatente alvo);
     void statusParcial();
