+++ tarefa4/src/com/rpg/personagens/CavaleiroCorrompido.java
@@ -20,9 +20,9 @@ public class CavaleiroCorrompido extends Monstro {
 
     @Override
     public AcaoDeCombate escolherAcao(Combatente alvo){
-        int distancia = Utilidades.calcularDistancia(pos, alvo.getPos());
+        double distancia = Utilidades.calcularDistancia(pos, alvo.getPos());
 
-        System.out.printf("\nO Cavaleiro Corrompido esta a %d metro(s) do heroi e ira ", distancia);
+        System.out.printf("\nO Cavaleiro Corrompido esta a %.0f metro(s) do heroi e ira ", distancia);
 
         if (distancia <= arma.getAttackRange()){
             System.out.println("ATACAR!\n"); Utilidades.esperar();
