+++ tarefa4/src/com/rpg/personagens/Zumbi.java
@@ -22,9 +22,9 @@ public class Zumbi extends Monstro {
 
     @Override
     public AcaoDeCombate escolherAcao(Combatente alvo){
-        int distancia = Utilidades.calcularDistancia(pos, alvo.getPos());
+        double distancia = Utilidades.calcularDistancia(pos, alvo.getPos());
 
-        System.out.printf("\nO ZUMBI está a %d metro(s) do herói e irá ", distancia);
+        System.out.printf("\nO ZUMBI está a %.0f metro(s) do herói e irá ", distancia);
 
         if (distancia <= arma.getAttackRange()){
             System.out.println("ATACAR!\n\n"); Utilidades.esperar();
